stored theme in localStorage
